fix: restructure useWrapCallbackHook
